// src/app/modules/schedule-edit/dialogs/schedule-add-genral-lesson-dialog/schedule-add-general-lesson-dialog.dto.ts
export interface ScheduleAddGeneralLessonFormData {
  subjectId?: string | null;
  teacherId?: string | null;
  groupId?: string | null;
  roomId?: string | null;
  primaryColor?: string | null;
  secondaryColor?: string | null;
  lessonIndex?: number | null;
  dayIndex?: number | null;
  startTime?: number | null;
  endTime?: number | null;
}